HAL library: split into manageable source files
This became just too unwieldy. Split source into one-per-HAL object,
and related support functions, so at least all code pertaining to a
particular HAL object can be found in a single short file.

Add hal_internal.h for intrs-library symbol resolution.

Move hal_data_u to the public API, it is just too useful to remain hidden.

Done in a single commit to minimally break git history.
