+++ src/hal/lib/Submakefile
@@ -13,6 +13,17 @@ HALLIBSRCS := $(HALLIBDIR)/hal_lib.c \
 	$(HALLIBDIR)/hal_iter.c \
 	$(HALLIBDIR)/hal_rcomp.c \
 	$(HALLIBDIR)/hal_vtable.c \
+	$(HALLIBDIR)/hal_funct.c \
+	$(HALLIBDIR)/hal_procfs.c \
+	$(HALLIBDIR)/hal_thread.c \
+	$(HALLIBDIR)/hal_param.c \
+	$(HALLIBDIR)/hal_alias.c \
+	$(HALLIBDIR)/hal_signal.c \
+	$(HALLIBDIR)/hal_pin.c \
+	$(HALLIBDIR)/hal_comp.c \
+	$(HALLIBDIR)/hal_list.c \
+	$(HALLIBDIR)/hal_memory.c \
+	$(HALLIBDIR)/hal_misc.c \
 	rtapi/rtapi_heap.c
 
 # protobuf support functions which depend on HAL - on RT host only
@@ -72,5 +83,16 @@ hal_lib-objs += hal/lib/hal_group.o
 hal_lib-objs += hal/lib/hal_ring.o
 hal_lib-objs += hal/lib/hal_rcomp.o
 hal_lib-objs += hal/lib/hal_vtable.o
+hal_lib-objs += hal/lib/hal_funct.o
+hal_lib-objs += hal/lib/hal_procfs.o
+hal_lib-objs += hal/lib/hal_thread.o
+hal_lib-objs += hal/lib/hal_signal.o
+hal_lib-objs += hal/lib/hal_pin.o
+hal_lib-objs += hal/lib/hal_param.o
+hal_lib-objs += hal/lib/hal_alias.o
+hal_lib-objs += hal/lib/hal_comp.o
+hal_lib-objs += hal/lib/hal_list.o
+hal_lib-objs += hal/lib/hal_memory.o
+hal_lib-objs += hal/lib/hal_misc.o
 
 $(RTLIBDIR)/hal_lib$(MODULE_EXT): $(addprefix $(OBJDIR)/,$(hal_lib-objs))
