+++ src/hal/lib/hal.h
@@ -468,6 +468,17 @@ typedef double real_t __attribute__((aligned(8)));
 typedef __u64 ireal_t __attribute__((aligned(8))); // integral type as wide as real_t / hal_float_t
 #define hal_float_t volatile real_t
 
+
+/** HAL "data union" structure
+ ** This structure may hold any type of hal data
+*/
+typedef union {
+    hal_bit_t b;
+    hal_s32_t s;
+    hal_s32_t u;
+    hal_float_t f;
+} hal_data_u;
+
 // type tags of HAL objects. See also protobuf/proto/types.proto/enum ObjectType
 // which must match:
 typedef enum {
