+++ src/hal/lib/hal_lib.c
@@ -63,6 +63,7 @@
 #include "hal_priv.h"		/* HAL private decls */
 #include "hal_ring.h"		/* HAL ringbuffer decls */
 #include "hal_group.h"		/* HAL group decls */
+#include "hal_internal.h"
 
 #include "rtapi_string.h"
 #ifdef RTAPI
@@ -113,14 +114,6 @@ static int lib_mem_id = -1;	/* RTAPI shmem ID for library module */
 // defined(BUILD_SYS_USER_DSO) && defined(RTAPI)
 // defined & attached in, and exported from rtapi_main.c
 
-// ULAPI: defined & attached in, & exported from ulapi_autoload.c
-extern global_data_t *global_data;
-
-// this is the pointer through which _all_ RTAPI/ULAPI references go:
-extern rtapi_switch_t *rtapi_switch;
-
-extern void  rtapi_printall(void);
-
 /***********************************************************************
 *                  LOCAL FUNCTION DECLARATIONS                         *
 ************************************************************************/
@@ -145,315 +138,11 @@ static int init_hal_data(void);
     All of these functions assume that the caller has already
     grabbed the hal_data mutex.
 */
-hal_comp_t *halpr_alloc_comp_struct(void);
-static hal_pin_t *alloc_pin_struct(void);
-static hal_sig_t *alloc_sig_struct(void);
-static hal_param_t *alloc_param_struct(void);
 
-static hal_oldname_t *halpr_alloc_oldname_struct(void);
-#ifdef RTAPI
-static hal_funct_t *alloc_funct_struct(void);
-#endif /* RTAPI */
-static hal_funct_entry_t *alloc_funct_entry_struct(void);
-#ifdef RTAPI
-static hal_thread_t *alloc_thread_struct(void);
-#endif /* RTAPI */
-
-static void free_comp_struct(hal_comp_t * comp);
-static void unlink_pin(hal_pin_t * pin);
-static void free_pin_struct(hal_pin_t * pin);
-static void free_sig_struct(hal_sig_t * sig);
-static void free_param_struct(hal_param_t * param);
-static void free_oldname_struct(hal_oldname_t * oldname);
-#ifdef RTAPI
-static void free_funct_struct(hal_funct_t * funct);
-#endif /* RTAPI */
-static void free_funct_entry_struct(hal_funct_entry_t * funct_entry);
-#ifdef RTAPI
-static void free_thread_struct(hal_thread_t * thread);
-#endif /* RTAPI */
-
-#ifdef RTAPI
-/** 'thread_task()' is a function that is invoked as a realtime task.
-    It implements a thread, by running down the thread's function list
-    and calling each function in turn.
-*/
-static void thread_task(void *arg);
-#endif /* RTAPI */
 
 /***********************************************************************
 *                  PUBLIC (API) FUNCTION CODE                          *
 ************************************************************************/
-int hal_init_mode(const char *name, int type, int userarg1, int userarg2)
-{
-    int comp_id;
-    char rtapi_name[RTAPI_NAME_LEN + 1];
-    char hal_name[HAL_NAME_LEN + 1];
-
-    // tag message origin field
-    rtapi_set_logtag("hal_lib");
-
-    if (name == 0) {
-	hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: no component name\n");
-	return -EINVAL;
-    }
-    if (strlen(name) > HAL_NAME_LEN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: component name '%s' is too long\n", name);
-	return -EINVAL;
-    }
-    // rtapi initialisation already done
-    // since this happens through the constructor
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: initializing component '%s' type=%d arg1=%d arg2=%d/0x%x\n",
-		    name, type, userarg1, userarg2, userarg2);
-    /* copy name to local vars, truncating if needed */
-    rtapi_snprintf(rtapi_name, RTAPI_NAME_LEN, "HAL_%s", name);
-    rtapi_snprintf(hal_name, sizeof(hal_name), "%s", name);
-
-    /* do RTAPI init */
-    comp_id = rtapi_init(rtapi_name);
-    if (comp_id < 0) {
-	hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: rtapi init failed\n");
-	return -EINVAL;
-    }
-    // tag message origin field since ulapi autoload re-tagged them
-    rtapi_set_logtag("hal_lib");
-#ifdef ULAPI
-    hal_rtapi_attach();
-#endif
-    {
-	hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before manipulating the shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-	/* make sure name is unique in the system */
-	if (halpr_find_comp_by_name(hal_name) != 0) {
-	    /* a component with this name already exists */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: duplicate component name '%s'\n", hal_name);
-	    rtapi_exit(comp_id);
-	    return -EINVAL;
-	}
-	/* allocate a new component structure */
-	comp = halpr_alloc_comp_struct();
-	if (comp == 0) {
-	    /* couldn't allocate structure */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for component '%s'\n", hal_name);
-	    rtapi_exit(comp_id);
-	    return -ENOMEM;
-	}
-
-	/* initialize the comp structure */
-	comp->userarg1 = userarg1;
-	comp->userarg2 = userarg2;
-	comp->comp_id = comp_id;
-	comp->type = type;
-#ifdef RTAPI
-	comp->pid = 0;   //FIXME revisit this
-#else /* ULAPI */
-	// a remote component starts out disowned
-	comp->pid = comp->type == TYPE_REMOTE ? 0 : getpid(); //FIXME revisit this
-#endif
-	comp->state = COMP_INITIALIZING;
-	comp->last_update = 0;
-	comp->last_bound = 0;
-	comp->last_unbound = 0;
-	comp->shmem_base = hal_shmem_base;
-	comp->insmod_args = 0;
-	rtapi_snprintf(comp->name, sizeof(comp->name), "%s", hal_name);
-	/* insert new structure at head of list */
-	comp->next_ptr = hal_data->comp_list_ptr;
-	hal_data->comp_list_ptr = SHMOFF(comp);
-
-    }
-    // scope exited - mutex released
-    /* done */
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: component '%s' initialized, ID = %02d\n", hal_name, comp_id);
-    return comp_id;
-}
-
-
-int hal_exit(int comp_id)
-{
-    int *prev, next;
-    char name[HAL_NAME_LEN + 1];
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: exit called before init\n");
-	return -EINVAL;
-    }
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: removing component %02d\n", comp_id);
-
-    {
-	hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* grab mutex before manipulating list */
-	rtapi_mutex_get(&(hal_data->mutex));
-	/* search component list for 'comp_id' */
-	prev = &(hal_data->comp_list_ptr);
-	next = *prev;
-	if (next == 0) {
-	    /* list is empty - should never happen, but... */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: component %d not found\n", comp_id);
-	    return -EINVAL;
-	}
-	comp = SHMPTR(next);
-	while (comp->comp_id != comp_id) {
-	    /* not a match, try the next one */
-	    prev = &(comp->next_ptr);
-	    next = *prev;
-	    if (next == 0) {
-		/* reached end of list without finding component */
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: component %d not found\n", comp_id);
-		return -EINVAL;
-	    }
-	    comp = SHMPTR(next);
-	}
-	/* found our component, unlink it from the list */
-	*prev = comp->next_ptr;
-	/* save component name for later */
-	rtapi_snprintf(name, sizeof(name), "%s", comp->name);
-	/* get rid of the component */
-	free_comp_struct(comp);
-	/*! \todo Another #if 0 */
-#if 0
-	/*! \todo FIXME - this is the beginning of a two pronged approach to managing
-	  shared memory.  Prong 1 - re-init the shared memory allocator whenever
-	  it is known to be safe.  Prong 2 - make a better allocator that can
-	  reclaim memory allocated by components when those components are
-	  removed. To be finished later. */
-	/* was that the last component? */
-	if (hal_data->comp_list_ptr == 0) {
-	    /* yes, are there any signals or threads defined? */
-	    if ((hal_data->sig_list_ptr == 0) && (hal_data->thread_list_ptr == 0)) {
-		/* no, invalidate "magic" number so shmem will be re-inited when
-		   a new component is loaded */
-		hal_data->magic = 0;
-	    }
-	}
-#endif
-	// scope exit - mutex released
-    }
-    // the RTAPI resources are now released
-    // on hal_lib shared library unload
-    rtapi_exit(comp_id);
-    /* done */
-    hal_print_msg(RTAPI_MSG_DBG,
-	"HAL: component %02d removed, name = '%s'\n", comp_id, name);
-
-    return 0;
-}
-
-void *hal_malloc(long int size)
-{
-    void *retval;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: hal_malloc called before init\n");
-	return 0;
-    }
-    /* get the mutex */
-    rtapi_mutex_get(&(hal_data->mutex));
-    /* allocate memory */
-    retval = shmalloc_up(size);
-    /* release the mutex */
-    rtapi_mutex_give(&(hal_data->mutex));
-    /* check return value */
-    if (retval == 0) {
-	hal_print_msg(RTAPI_MSG_DBG,
-	    "HAL: hal_malloc() can't allocate %ld bytes\n", size);
-    }
-    return retval;
-}
-
-#ifdef RTAPI
-int hal_set_constructor(int comp_id, constructor make) {
-    int next;
-    hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-    rtapi_mutex_get(&(hal_data->mutex));
-
-    /* search component list for 'comp_id' */
-    next = hal_data->comp_list_ptr;
-    if (next == 0) {
-	/* list is empty - should never happen, but... */
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: component %d not found\n", comp_id);
-	return -EINVAL;
-    }
-
-    comp = SHMPTR(next);
-    while (comp->comp_id != comp_id) {
-	/* not a match, try the next one */
-	next = comp->next_ptr;
-	if (next == 0) {
-	    /* reached end of list without finding component */
-	    hal_print_msg(RTAPI_MSG_ERR,
-		"HAL: ERROR: component %d not found\n", comp_id);
-	    return -EINVAL;
-	}
-	comp = SHMPTR(next);
-    }
-    comp->make = make;
-    return 0;
-}
-#endif
-
-int hal_ready(int comp_id) {
-    int next;
-    hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-    rtapi_mutex_get(&(hal_data->mutex));
-
-    /* search component list for 'comp_id' */
-    next = hal_data->comp_list_ptr;
-    if (next == 0) {
-	/* list is empty - should never happen, but... */
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: component %d not found\n", comp_id);
-	return -EINVAL;
-    }
-
-    comp = SHMPTR(next);
-    while (comp->comp_id != comp_id) {
-	/* not a match, try the next one */
-	next = comp->next_ptr;
-	if (next == 0) {
-	    /* reached end of list without finding component */
-	    hal_print_msg(RTAPI_MSG_ERR,
-		"HAL: ERROR: component %d not found\n", comp_id);
-	    return -EINVAL;
-	}
-	comp = SHMPTR(next);
-    }
-    if(comp->state > COMP_INITIALIZING) {
-        hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: Component '%s' already ready (%d)\n",
-			comp->name, comp->state);
-        return -EINVAL;
-    }
-    comp->state = (comp->type == TYPE_REMOTE ?  COMP_UNBOUND : COMP_READY);
-    return 0;
-}
-
-char *hal_comp_name(int comp_id)
-{
-    hal_comp_t *comp;
-    char *result = NULL;
-    rtapi_mutex_get(&(hal_data->mutex));
-    comp = halpr_find_comp_by_id(comp_id);
-    if(comp) result = comp->name;
-    rtapi_mutex_give(&(hal_data->mutex));
-    return result;
-}
-
 /***********************************************************************
 *                      "LOCKING" FUNCTIONS                             *
 ************************************************************************/
@@ -483,2072 +172,6 @@ unsigned char hal_get_lock() {
     return hal_data->lock;
 }
 
-
-/***********************************************************************
-*                        "PIN" FUNCTIONS                               *
-************************************************************************/
-
-/* wrapper functs for typed pins - these call the generic funct below */
-
-int hal_pin_bit_new(const char *name, hal_pin_dir_t dir,
-    hal_bit_t ** data_ptr_addr, int comp_id)
-{
-    return hal_pin_new(name, HAL_BIT, dir, (void **) data_ptr_addr, comp_id);
-}
-
-int hal_pin_float_new(const char *name, hal_pin_dir_t dir,
-    hal_float_t ** data_ptr_addr, int comp_id)
-{
-    return hal_pin_new(name, HAL_FLOAT, dir, (void **) data_ptr_addr,
-	comp_id);
-}
-
-int hal_pin_u32_new(const char *name, hal_pin_dir_t dir,
-    hal_u32_t ** data_ptr_addr, int comp_id)
-{
-    return hal_pin_new(name, HAL_U32, dir, (void **) data_ptr_addr, comp_id);
-}
-
-int hal_pin_s32_new(const char *name, hal_pin_dir_t dir,
-    hal_s32_t ** data_ptr_addr, int comp_id)
-{
-    return hal_pin_new(name, HAL_S32, dir, (void **) data_ptr_addr, comp_id);
-}
-
-static int hal_pin_newfv(hal_type_t type, hal_pin_dir_t dir,
-    void ** data_ptr_addr, int comp_id, const char *fmt, va_list ap)
-{
-    char name[HAL_NAME_LEN + 1];
-    int sz;
-    sz = rtapi_vsnprintf(name, sizeof(name), fmt, ap);
-    if(sz == -1 || sz > HAL_NAME_LEN) {
-        hal_print_msg(RTAPI_MSG_ERR,
-	    "hal_pin_newfv: length %d too long for name starting '%s'\n",
-	    sz, name);
-        return -ENOMEM;
-    }
-    return hal_pin_new(name, type, dir, data_ptr_addr, comp_id);
-}
-
-int hal_pin_bit_newf(hal_pin_dir_t dir,
-    hal_bit_t ** data_ptr_addr, int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_pin_newfv(HAL_BIT, dir, (void**)data_ptr_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-int hal_pin_float_newf(hal_pin_dir_t dir,
-    hal_float_t ** data_ptr_addr, int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_pin_newfv(HAL_FLOAT, dir, (void**)data_ptr_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-int hal_pin_u32_newf(hal_pin_dir_t dir,
-    hal_u32_t ** data_ptr_addr, int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_pin_newfv(HAL_U32, dir, (void**)data_ptr_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-int hal_pin_s32_newf(hal_pin_dir_t dir,
-    hal_s32_t ** data_ptr_addr, int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_pin_newfv(HAL_S32, dir, (void**)data_ptr_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-
-/* this is a generic function that does the majority of the work. */
-
-int hal_pin_new(const char *name, hal_type_t type, hal_pin_dir_t dir,
-    void **data_ptr_addr, int comp_id)
-{
-    int *prev, next, cmp;
-    hal_pin_t *new, *ptr;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin_new called before init\n");
-	return -EINVAL;
-    }
-
-    if(*data_ptr_addr)
-    {
-        hal_print_msg(RTAPI_MSG_ERR,
-            "HAL: ERROR: pin_new(%s) called with already-initialized memory\n",
-            name);
-    }
-    if (type != HAL_BIT && type != HAL_FLOAT && type != HAL_S32 && type != HAL_U32) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin type not one of HAL_BIT, HAL_FLOAT, HAL_S32 or HAL_U32\n");
-	return -EINVAL;
-    }
-
-    if(dir != HAL_IN && dir != HAL_OUT && dir != HAL_IO) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin direction not one of HAL_IN, HAL_OUT, or HAL_IO\n");
-	return -EINVAL;
-    }
-    if (strlen(name) > HAL_NAME_LEN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin name '%s' is too long\n", name);
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_LOAD)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin_new called while HAL locked\n");
-	return -EPERM;
-    }
-
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: creating pin '%s'\n", name);
-
-    {
-	hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	/* validate comp_id */
-	comp = halpr_find_comp_by_id(comp_id);
-	if (comp == 0) {
-	    /* bad comp_id */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: component %d not found\n", comp_id);
-	    return -EINVAL;
-	}
-	/* validate passed in pointer - must point to HAL shmem */
-	if (! SHMCHK(data_ptr_addr)) {
-	    /* bad pointer */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: data_ptr_addr not in shared memory\n");
-	    return -EINVAL;
-	}
-	if(comp->state > COMP_INITIALIZING) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: pin_new called after hal_ready (%d)\n", comp->state);
-	    return -EINVAL;
-	}
-
-	/* allocate a new variable structure */
-	new = alloc_pin_struct();
-	if (new == 0) {
-	    /* alloc failed */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for pin '%s'\n", name);
-	    return -ENOMEM;
-	}
-	/* initialize the structure */
-	new->data_ptr_addr = SHMOFF(data_ptr_addr);
-	new->owner_ptr = SHMOFF(comp);
-	new->type = type;
-	new->dir = dir;
-	new->signal = 0;
-	new->handle = rtapi_next_handle();
-	memset(&new->dummysig, 0, sizeof(hal_data_u));
-	rtapi_snprintf(new->name, sizeof(new->name), "%s", name);
-	/* make 'data_ptr' point to dummy signal */
-	*data_ptr_addr = comp->shmem_base + SHMOFF(&(new->dummysig));
-	/* search list for 'name' and insert new structure */
-	prev = &(hal_data->pin_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		return 0;
-	    }
-	    ptr = SHMPTR(next);
-	    cmp = strcmp(ptr->name, new->name);
-	    if (cmp > 0) {
-		/* found the right place for it, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		return 0;
-	    }
-	    if (cmp == 0) {
-		/* name already in list, can't insert */
-		free_pin_struct(new);
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: duplicate variable '%s'\n", name);
-		return -EINVAL;
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(ptr->next_ptr);
-	    next = *prev;
-	}
-    }
-}
-
-int hal_pin_alias(const char *pin_name, const char *alias)
-{
-    int *prev, next, cmp;
-    hal_pin_t *pin, *ptr;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin_alias called before init\n");
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_CONFIG)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: pin_alias called while HAL locked\n");
-	return -EPERM;
-    }
-    if (alias != NULL ) {
-	if (strlen(alias) > HAL_NAME_LEN) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-	        "HAL: ERROR: alias name '%s' is too long\n", alias);
-	    return -EINVAL;
-	}
-    }
-    {
-	hal_oldname_t *oldname  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-	if (alias != NULL ) {
-	    pin = halpr_find_pin_by_name(alias);
-	    if ( pin != NULL ) {
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: duplicate pin/alias name '%s'\n",
-				alias);
-		return -EINVAL;
-	    }
-	}
-	/* once we unlink the pin from the list, we don't want to have to
-	   abort the change and repair things.  So we allocate an oldname
-	   struct here, then free it (which puts it on the free list).  This
-	   allocation might fail, in which case we abort the command.  But
-	   if we actually need the struct later, the next alloc is guaranteed
-	   to succeed since at least one struct is on the free list. */
-	oldname = halpr_alloc_oldname_struct();
-	if ( oldname == NULL ) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for pin_alias\n");
-	    return -EINVAL;
-	}
-	free_oldname_struct(oldname);
-
-	/* find the pin and unlink it from pin list */
-	prev = &(hal_data->pin_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, not found */
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: pin '%s' not found\n", pin_name);
-		return -EINVAL;
-	    }
-	    pin = SHMPTR(next);
-	    if ( strcmp(pin->name, pin_name) == 0 ) {
-		/* found it, unlink from list */
-		*prev = pin->next_ptr;
-		break;
-	    }
-	    if (pin->oldname != 0 ) {
-		oldname = SHMPTR(pin->oldname);
-		if (strcmp(oldname->name, pin_name) == 0) {
-		    /* found it, unlink from list */
-		    *prev = pin->next_ptr;
-		    break;
-		}
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(pin->next_ptr);
-	    next = *prev;
-	}
-	if ( alias != NULL ) {
-	/* adding a new alias */
-	    if ( pin->oldname == 0 ) {
-		/* save old name (only if not already saved) */
-		oldname = halpr_alloc_oldname_struct();
-		pin->oldname = SHMOFF(oldname);
-		rtapi_snprintf(oldname->name, sizeof(oldname->name),
-			       "%s", pin->name);
-	    }
-	    /* change pin's name to 'alias' */
-	    rtapi_snprintf(pin->name, sizeof(pin->name), "%s", alias);
-	} else {
-	    /* removing an alias */
-	    if ( pin->oldname != 0 ) {
-		/* restore old name (only if pin is aliased) */
-	    oldname = SHMPTR(pin->oldname);
-	    rtapi_snprintf(pin->name, sizeof(pin->name), "%s", oldname->name);
-	    pin->oldname = 0;
-	    free_oldname_struct(oldname);
-	    }
-	}
-	/* insert pin back into list in proper place */
-	prev = &(hal_data->pin_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, insert here */
-		pin->next_ptr = next;
-		*prev = SHMOFF(pin);
-		return 0;
-	    }
-	    ptr = SHMPTR(next);
-	    cmp = strcmp(ptr->name, pin->name);
-	    if (cmp > 0) {
-		/* found the right place for it, insert here */
-		pin->next_ptr = next;
-		*prev = SHMOFF(pin);
-		return 0;
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(ptr->next_ptr);
-	    next = *prev;
-	}
-    }
-}
-/***********************************************************************
-*                      "SIGNAL" FUNCTIONS                              *
-************************************************************************/
-
-int hal_signal_new(const char *name, hal_type_t type)
-{
-
-    int *prev, next, cmp;
-    hal_sig_t *new, *ptr;
-
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: signal_new called before init\n");
-	return -EINVAL;
-    }
-
-    if (strlen(name) > HAL_NAME_LEN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: signal name '%s' is too long\n", name);
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_CONFIG) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: signal_new called while HAL is locked\n");
-	return -EPERM;
-    }
-    {
-	void *data_addr  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	hal_print_msg(RTAPI_MSG_DBG, "HAL: creating signal '%s'\n", name);
-
-	/* check for an existing signal with the same name */
-	if (halpr_find_sig_by_name(name) != 0) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: duplicate signal '%s'\n", name);
-	    return -EINVAL;
-	}
-	/* allocate memory for the signal value */
-	switch (type) {
-	case HAL_BIT:
-	    data_addr = shmalloc_up(sizeof(hal_bit_t));
-	    break;
-	case HAL_S32:
-	    data_addr = shmalloc_up(sizeof(hal_u32_t));
-	    break;
-	case HAL_U32:
-	    data_addr = shmalloc_up(sizeof(hal_s32_t));
-	    break;
-	case HAL_FLOAT:
-	    data_addr = shmalloc_up(sizeof(hal_float_t));
-	    break;
-	default:
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: illegal signal type %d'\n", type);
-	    return -EINVAL;
-	    break;
-	}
-	/* allocate a new signal structure */
-	new = alloc_sig_struct();
-	if ((new == 0) || (data_addr == 0)) {
-	    /* alloc failed */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for signal '%s'\n",
-			    name);
-	    return -ENOMEM;
-	}
-	/* initialize the signal value */
-	switch (type) {
-	case HAL_BIT:
-	    *((char *) data_addr) = 0;
-	    break;
-	case HAL_S32:
-	    *((hal_s32_t *) data_addr) = 0;
-	    break;
-	case HAL_U32:
-	    *((hal_u32_t *) data_addr) = 0;
-	    break;
-	case HAL_FLOAT:
-	    *((hal_float_t *) data_addr) = 0.0;
-	    break;
-	default:
-	    break;
-	}
-	/* initialize the structure */
-	new->data_ptr = SHMOFF(data_addr);
-	new->type = type;
-	new->readers = 0;
-	new->writers = 0;
-	new->bidirs = 0;
-	new->handle = rtapi_next_handle();
-	rtapi_snprintf(new->name, sizeof(new->name), "%s", name);
-	/* search list for 'name' and insert new structure */
-	prev = &(hal_data->sig_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		return 0;
-	    }
-	    ptr = SHMPTR(next);
-	    cmp = strcmp(ptr->name, new->name);
-	    if (cmp > 0) {
-		/* found the right place for it, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		return 0;
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(ptr->next_ptr);
-	    next = *prev;
-	}
-    }
-}
-
-int hal_signal_delete(const char *name)
-{
-
-    int *prev, next;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: signal_delete called before init\n");
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_CONFIG)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: signal_delete called while HAL locked\n");
-	return -EPERM;
-    }
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: deleting signal '%s'\n", name);
-
-    {
-	hal_sig_t *sig  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-	/* search for the signal */
-	prev = &(hal_data->sig_list_ptr);
-	next = *prev;
-	while (next != 0) {
-	    sig = SHMPTR(next);
-	    if (strcmp(sig->name, name) == 0) {
-		hal_group_t *grp = halpr_find_group_of_member(name);
-		if (grp) {
-		    hal_print_msg(RTAPI_MSG_ERR,
-				    "HAL: ERROR: cannot delete signal '%s'"
-				    " since it is member of group '%s'\n",
-				    name, grp->name);
-		    return -EINVAL;
-		}
-		/* this is the right signal, unlink from list */
-		*prev = sig->next_ptr;
-		/* and delete it */
-		free_sig_struct(sig);
-		/* done */
-		return 0;
-	    }
-	    /* no match, try the next one */
-	    prev = &(sig->next_ptr);
-	    next = *prev;
-	}
-    }
-    /* if we get here, we didn't find a match */
-    hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: signal '%s' not found\n",
-		    name);
-    return -EINVAL;
-}
-
-
-int hal_link(const char *pin_name, const char *sig_name)
-{
-    hal_sig_t *sig;
-    hal_comp_t *comp;
-    void **data_ptr_addr, *data_addr;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: link called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_CONFIG)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: link called while HAL locked\n");
-	return -EPERM;
-    }
-    /* make sure we were given a pin name */
-    if (pin_name == 0) {
-	hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: pin name not given\n");
-	return -EINVAL;
-    }
-    /* make sure we were given a signal name */
-    if (sig_name == 0) {
-	hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: signal name not given\n");
-	return -EINVAL;
-    }
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: linking pin '%s' to '%s'\n", pin_name, sig_name);
-
-    {
-	hal_pin_t *pin  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing data structures */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	/* locate the pin */
-	pin = halpr_find_pin_by_name(pin_name);
-	if (pin == 0) {
-	    /* not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: pin '%s' not found\n", pin_name);
-	    return -EINVAL;
-	}
-	/* locate the signal */
-	sig = halpr_find_sig_by_name(sig_name);
-	if (sig == 0) {
-	    /* not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: signal '%s' not found\n", sig_name);
-	    return -EINVAL;
-	}
-	/* found both pin and signal, are they already connected? */
-	if (SHMPTR(pin->signal) == sig) {
-	    hal_print_msg(RTAPI_MSG_WARN,
-			    "HAL: Warning: pin '%s' already linked to '%s'\n", pin_name, sig_name);
-	    return 0;
-	}
-	/* is the pin connected to something else? */
-	if(pin->signal) {
-	    sig = SHMPTR(pin->signal);
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: pin '%s' is linked to '%s', cannot link to '%s'\n",
-			    pin_name, sig->name, sig_name);
-	    return -EINVAL;
-	}
-	/* check types */
-	if (pin->type != sig->type) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: type mismatch '%s' <- '%s'\n", pin_name, sig_name);
-	    return -EINVAL;
-	}
-	/* linking output pin to sig that already has output or I/O pins? */
-	if ((pin->dir == HAL_OUT) && ((sig->writers > 0) || (sig->bidirs > 0 ))) {
-	    /* yes, can't do that */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: signal '%s' already has output or I/O pin(s)\n", sig_name);
-	    return -EINVAL;
-	}
-	/* linking bidir pin to sig that already has output pin? */
-	if ((pin->dir == HAL_IO) && (sig->writers > 0)) {
-	    /* yes, can't do that */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: signal '%s' already has output pin\n", sig_name);
-	    return -EINVAL;
-	}
-	/* everything is OK, make the new link */
-	data_ptr_addr = SHMPTR(pin->data_ptr_addr);
-	comp = SHMPTR(pin->owner_ptr);
-	data_addr = comp->shmem_base + sig->data_ptr;
-	*data_ptr_addr = data_addr;
-
-	if (( sig->readers == 0 ) && ( sig->writers == 0 ) && ( sig->bidirs == 0 )) {
-	    /* this is the first pin for this signal, copy value from pin's "dummy" field */
-	    data_addr = hal_shmem_base + sig->data_ptr;
-	    *((hal_data_u *)data_addr) = pin->dummysig;
-	}
-
-	/* update the signal's reader/writer/bidir counts */
-	if ((pin->dir & HAL_IN) != 0) {
-	    sig->readers++;
-	}
-	if (pin->dir == HAL_OUT) {
-	    sig->writers++;
-	}
-	if (pin->dir == HAL_IO) {
-	    sig->bidirs++;
-	}
-	/* and update the pin */
-	pin->signal = SHMOFF(sig);
-    }
-    return 0;
-}
-
-int hal_unlink(const char *pin_name)
-{
-
-
-    // preliminary checks are done before locking
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: unlink called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_CONFIG)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: unlink called while HAL locked\n");
-	return -EPERM;
-    }
-    /* make sure we were given a pin name */
-    if (pin_name == 0) {
-	hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: pin name not given\n");
-	return -EINVAL;
-    }
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: unlinking pin '%s'\n", pin_name);
-
-    {
-	hal_pin_t *pin __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing data structures */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	/* locate the pin */
-	pin = halpr_find_pin_by_name(pin_name);
-
-	if (pin == 0) {
-	    /* not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: pin '%s' not found\n", pin_name);
-	    return -EINVAL;
-	}
-
-	/* found pin, unlink it */
-	unlink_pin(pin);
-
-	// done, mutex will be released on scope exit automagically
-	return 0;
-    }
-}
-
-/***********************************************************************
-*                       "PARAM" FUNCTIONS                              *
-************************************************************************/
-
-/* wrapper functs for typed params - these call the generic funct below */
-
-int hal_param_bit_new(const char *name, hal_param_dir_t dir, hal_bit_t * data_addr,
-    int comp_id)
-{
-    return hal_param_new(name, HAL_BIT, dir, (void *) data_addr, comp_id);
-}
-
-int hal_param_float_new(const char *name, hal_param_dir_t dir, hal_float_t * data_addr,
-    int comp_id)
-{
-    return hal_param_new(name, HAL_FLOAT, dir, (void *) data_addr, comp_id);
-}
-
-int hal_param_u32_new(const char *name, hal_param_dir_t dir, hal_u32_t * data_addr,
-    int comp_id)
-{
-    return hal_param_new(name, HAL_U32, dir, (void *) data_addr, comp_id);
-}
-
-int hal_param_s32_new(const char *name, hal_param_dir_t dir, hal_s32_t * data_addr,
-    int comp_id)
-{
-    return hal_param_new(name, HAL_S32, dir, (void *) data_addr, comp_id);
-}
-
-static int hal_param_newfv(hal_type_t type, hal_param_dir_t dir,
-	void *data_addr, int comp_id, const char *fmt, va_list ap) {
-    char name[HAL_NAME_LEN + 1];
-    int sz;
-    sz = rtapi_vsnprintf(name, sizeof(name), fmt, ap);
-    if(sz == -1 || sz > HAL_NAME_LEN) {
-        hal_print_msg(RTAPI_MSG_ERR,
-	    "hal_param_newfv: length %d too long for name starting '%s'\n",
-	    sz, name);
-	return -ENOMEM;
-    }
-    return hal_param_new(name, type, dir, (void *) data_addr, comp_id);
-}
-
-int hal_param_bit_newf(hal_param_dir_t dir, hal_bit_t * data_addr,
-    int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_param_newfv(HAL_BIT, dir, (void*)data_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-int hal_param_float_newf(hal_param_dir_t dir, hal_float_t * data_addr,
-    int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_param_newfv(HAL_FLOAT, dir, (void*)data_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-int hal_param_u32_newf(hal_param_dir_t dir, hal_u32_t * data_addr,
-    int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_param_newfv(HAL_U32, dir, (void*)data_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-int hal_param_s32_newf(hal_param_dir_t dir, hal_s32_t * data_addr,
-    int comp_id, const char *fmt, ...)
-{
-    va_list ap;
-    int ret;
-    va_start(ap, fmt);
-    ret = hal_param_newfv(HAL_S32, dir, (void*)data_addr, comp_id, fmt, ap);
-    va_end(ap);
-    return ret;
-}
-
-
-/* this is a generic function that does the majority of the work. */
-
-int hal_param_new(const char *name, hal_type_t type, hal_param_dir_t dir, void *data_addr,
-		  int comp_id)
-{
-    int *prev, next, cmp;
-    hal_param_t *new, *ptr;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param_new called before init\n");
-	return -EINVAL;
-    }
-
-    if (type != HAL_BIT && type != HAL_FLOAT && type != HAL_S32 && type != HAL_U32) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: pin type not one of HAL_BIT, HAL_FLOAT, HAL_S32 or HAL_U32\n");
-	return -EINVAL;
-    }
-
-    if (dir != HAL_RO && dir != HAL_RW) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param direction not one of HAL_RO, or HAL_RW\n");
-	return -EINVAL;
-    }
-
-    if (strlen(name) > HAL_NAME_LEN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: parameter name '%s' is too long\n", name);
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_LOAD)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param_new called while HAL locked\n");
-	return -EPERM;
-    }
-    {
-	hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	hal_print_msg(RTAPI_MSG_DBG, "HAL: creating parameter '%s'\n", name);
-
-	/* validate comp_id */
-	comp = halpr_find_comp_by_id(comp_id);
-	if (comp == 0) {
-	    /* bad comp_id */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: component %d not found\n", comp_id);
-	    return -EINVAL;
-	}
-	/* validate passed in pointer - must point to HAL shmem */
-	if (! SHMCHK(data_addr)) {
-	    /* bad pointer */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: data_addr not in shared memory\n");
-	    return -EINVAL;
-	}
-	if(comp->state > COMP_INITIALIZING) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: param_new called after hal_ready\n");
-	    return -EINVAL;
-	}
-	/* allocate a new parameter structure */
-	new = alloc_param_struct();
-	if (new == 0) {
-	    /* alloc failed */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for parameter '%s'\n", name);
-	    return -ENOMEM;
-	}
-	/* initialize the structure */
-	new->owner_ptr = SHMOFF(comp);
-	new->data_ptr = SHMOFF(data_addr);
-	new->type = type;
-	new->dir = dir;
-	new->handle = rtapi_next_handle();
-	rtapi_snprintf(new->name, sizeof(new->name), "%s", name);
-	/* search list for 'name' and insert new structure */
-	prev = &(hal_data->param_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		return 0;
-	    }
-	    ptr = SHMPTR(next);
-	    cmp = strcmp(ptr->name, new->name);
-	    if (cmp > 0) {
-		/* found the right place for it, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		return 0;
-	    }
-	    if (cmp == 0) {
-		/* name already in list, can't insert */
-		free_param_struct(new);
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: duplicate parameter '%s'\n", name);
-		return -EINVAL;
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(ptr->next_ptr);
-	    next = *prev;
-	}
-    }
-}
-
-/* wrapper functs for typed params - these call the generic funct below */
-
-int hal_param_bit_set(const char *name, int value)
-{
-    return hal_param_set(name, HAL_BIT, &value);
-}
-
-int hal_param_float_set(const char *name, double value)
-{
-    return hal_param_set(name, HAL_FLOAT, &value);
-}
-
-int hal_param_u32_set(const char *name, unsigned long value)
-{
-    return hal_param_set(name, HAL_U32, &value);
-}
-
-int hal_param_s32_set(const char *name, signed long value)
-{
-    return hal_param_set(name, HAL_S32, &value);
-}
-
-/* this is a generic function that does the majority of the work */
-
-int hal_param_set(const char *name, hal_type_t type, void *value_addr)
-{
-
-    void *d_ptr;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param_set called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_PARAMS)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param_set called while HAL locked\n");
-	return -EPERM;
-    }
-
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: setting parameter '%s'\n", name);
-
-    {
-	hal_param_t *param __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	/* search param list for name */
-	param = halpr_find_param_by_name(name);
-	if (param == 0) {
-	    /* parameter not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: parameter '%s' not found\n", name);
-	    return -EINVAL;
-	}
-	/* found it, is type compatible? */
-	if (param->type != type) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: type mismatch setting param '%s'\n",
-			    name);
-	    return -EINVAL;
-	}
-	/* is it read only? */
-	if (param->dir == HAL_RO) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: param '%s' is not writable\n",
-			    name);
-	    return -EINVAL;
-	}
-	/* everything is OK, set the value */
-	d_ptr = SHMPTR(param->data_ptr);
-	switch (param->type) {
-	case HAL_BIT:
-	    if (*((int *) value_addr) == 0) {
-		*(hal_bit_t *) (d_ptr) = 0;
-	    } else {
-		*(hal_bit_t *) (d_ptr) = 1;
-	    }
-	    break;
-	case HAL_FLOAT:
-	    *((hal_float_t *) (d_ptr)) = *((double *) (value_addr));
-	    break;
-	case HAL_S32:
-	    *((hal_s32_t *) (d_ptr)) = *((signed long *) (value_addr));
-	    break;
-	case HAL_U32:
-	    *((hal_u32_t *) (d_ptr)) = *((unsigned long *) (value_addr));
-	    break;
-	default:
-	    /* Shouldn't get here, but just in case... */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: bad type %d setting param\n", param->type);
-	    return -EINVAL;
-	}
-    }
-    return 0;
-}
-
-int hal_param_alias(const char *param_name, const char *alias)
-{
-    int *prev, next, cmp;
-    hal_param_t *param, *ptr;
-
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param_alias called before init\n");
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_CONFIG)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: param_alias called while HAL locked\n");
-	return -EPERM;
-    }
-    if (alias != NULL ) {
-	if (strlen(alias) > HAL_NAME_LEN) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: alias name '%s' is too long\n", alias);
-	    return -EINVAL;
-	}
-    }
-
-    {
-	hal_oldname_t *oldname __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	if (alias != NULL ) {
-	    param = halpr_find_param_by_name(alias);
-	    if ( param != NULL ) {
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: duplicate pin/alias name '%s'\n", alias);
-		return -EINVAL;
-	    }
-	}
-	/* once we unlink the param from the list, we don't want to have to
-	   abort the change and repair things.  So we allocate an oldname
-	   struct here, then free it (which puts it on the free list).  This
-	   allocation might fail, in which case we abort the command.  But
-	   if we actually need the struct later, the next alloc is guaranteed
-	   to succeed since at least one struct is on the free list. */
-	oldname = halpr_alloc_oldname_struct();
-	if ( oldname == NULL ) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for param_alias\n");
-	    return -EINVAL;
-	}
-	free_oldname_struct(oldname);
-	/* find the param and unlink it from pin list */
-	prev = &(hal_data->param_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, not found */
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: param '%s' not found\n", param_name);
-		return -EINVAL;
-	    }
-	    param = SHMPTR(next);
-	    if ( strcmp(param->name, param_name) == 0 ) {
-		/* found it, unlink from list */
-		*prev = param->next_ptr;
-		break;
-	    }
-	    if (param->oldname != 0 ) {
-		oldname = SHMPTR(param->oldname);
-		if (strcmp(oldname->name, param_name) == 0) {
-		    /* found it, unlink from list */
-		    *prev = param->next_ptr;
-		    break;
-		}
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(param->next_ptr);
-	    next = *prev;
-	}
-	if ( alias != NULL ) {
-	    /* adding a new alias */
-	    if ( param->oldname == 0 ) {
-		/* save old name (only if not already saved) */
-		oldname = halpr_alloc_oldname_struct();
-		param->oldname = SHMOFF(oldname);
-		rtapi_snprintf(oldname->name, sizeof(oldname->name), "%s", param->name);
-	    }
-	    /* change param's name to 'alias' */
-	    rtapi_snprintf(param->name, sizeof(param->name), "%s", alias);
-	} else {
-	    /* removing an alias */
-	    if ( param->oldname != 0 ) {
-		/* restore old name (only if param is aliased) */
-		oldname = SHMPTR(param->oldname);
-		rtapi_snprintf(param->name, sizeof(param->name), "%s", oldname->name);
-		param->oldname = 0;
-		free_oldname_struct(oldname);
-	    }
-	}
-	/* insert param back into list in proper place */
-	prev = &(hal_data->param_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, insert here */
-		param->next_ptr = next;
-		*prev = SHMOFF(param);
-		return 0;
-	    }
-	    ptr = SHMPTR(next);
-	    cmp = strcmp(ptr->name, param->name);
-	    if (cmp > 0) {
-		/* found the right place for it, insert here */
-		param->next_ptr = next;
-		*prev = SHMOFF(param);
-		return 0;
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(ptr->next_ptr);
-	    next = *prev;
-	}
-    }
-}
-
-/***********************************************************************
-*                   EXECUTION RELATED FUNCTIONS                        *
-************************************************************************/
-
-#ifdef RTAPI
-
-int hal_export_funct(const char *name, void (*funct) (void *, long),
-		     void *arg, int uses_fp, int reentrant, int comp_id)
-{
-    int *prev, next, cmp;
-    hal_funct_t *new, *fptr;
-    char buf[HAL_NAME_LEN + 1];
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: export_funct called before init\n");
-	return -EINVAL;
-    }
-
-    if (strlen(name) > HAL_NAME_LEN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: function name '%s' is too long\n", name);
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_LOAD)  {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: export_funct called while HAL locked\n");
-	return -EPERM;
-    }
-
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: exporting function '%s'\n", name);
-
-    {
-	hal_comp_t *comp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-	/* validate comp_id */
-	comp = halpr_find_comp_by_id(comp_id);
-	if (comp == 0) {
-	    /* bad comp_id */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: component %d not found\n", comp_id);
-	    return -EINVAL;
-	}
-	if (comp->type == TYPE_USER) {
-	    /* not a realtime component */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: component %d is not realtime (%d)\n",
-			    comp_id, comp->type);
-	    return -EINVAL;
-	}
-	if(comp->state > COMP_INITIALIZING) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: export_funct called after hal_ready\n");
-	    return -EINVAL;
-	}
-	/* allocate a new function structure */
-	new = alloc_funct_struct();
-	if (new == 0) {
-	    /* alloc failed */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for function '%s'\n", name);
-	    return -ENOMEM;
-	}
-	/* initialize the structure */
-	new->uses_fp = uses_fp;
-	new->owner_ptr = SHMOFF(comp);
-	new->reentrant = reentrant;
-	new->users = 0;
-	new->handle = rtapi_next_handle();
-	new->arg = arg;
-	new->funct = funct;
-	rtapi_snprintf(new->name, sizeof(new->name), "%s", name);
-	/* search list for 'name' and insert new structure */
-	prev = &(hal_data->funct_list_ptr);
-	next = *prev;
-	while (1) {
-	    if (next == 0) {
-		/* reached end of list, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		/* break out of loop and init the new function */
-		break;
-	    }
-	    fptr = SHMPTR(next);
-	    cmp = strcmp(fptr->name, new->name);
-	    if (cmp > 0) {
-		/* found the right place for it, insert here */
-		new->next_ptr = next;
-		*prev = SHMOFF(new);
-		/* break out of loop and init the new function */
-		break;
-	    }
-	    if (cmp == 0) {
-		/* name already in list, can't insert */
-		free_funct_struct(new);
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: duplicate function '%s'\n", name);
-		return -EINVAL;
-	    }
-	    /* didn't find it yet, look at next one */
-	    prev = &(fptr->next_ptr);
-	    next = *prev;
-	}
-	// at this point we have a new function and can
-	// yield the mutex by scope exit
-
-    }
-    /* init time logging variables */
-    new->runtime = 0;
-    new->maxtime = 0;
-    new->maxtime_increased = 0;
-
-    /* at this point we have a new function and can yield the mutex */
-    rtapi_mutex_give(&(hal_data->mutex));
-
-    /* create a pin with the function's runtime in it */
-    if (hal_pin_s32_newf(HAL_OUT, &(new->runtime), comp_id,"%s.time",name)) {
-	rtapi_print_msg(RTAPI_MSG_ERR,
-	   "HAL: ERROR: fail to create pin '%s.time'\n", name);
-	return -EINVAL;
-    }
-    *(new->runtime) = 0;
-
-    /* note that failure to successfully create the following params
-       does not cause the "export_funct()" call to fail - they are
-       for debugging and testing use only */
-    /* create a parameter with the function's maximum runtime in it */
-    rtapi_snprintf(buf, sizeof(buf), "%s.tmax", name);
-    new->maxtime = 0;
-    hal_param_s32_new(buf, HAL_RW, &(new->maxtime), comp_id);
-
-    /* create a parameter with the function's maximum runtime in it */
-    rtapi_snprintf(buf, sizeof(buf), "%s.tmax-increased", name);
-    new->maxtime_increased = 0;
-    hal_param_bit_new(buf, HAL_RO, &(new->maxtime_increased), comp_id);
-
-    return 0;
-}
-
-int hal_create_thread(const char *name, unsigned long period_nsec,
-		      int uses_fp, int cpu_id)
-{
-    int next, prev_priority;
-    int retval, n;
-    hal_thread_t *new, *tptr;
-    long prev_period, curr_period;
-    /*! \todo Another #if 0 */
-#if 0
-    char buf[HAL_NAME_LEN + 1];
-#endif
-
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: creating thread %s, %ld nsec\n", name, period_nsec);
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: create_thread called before init\n");
-	return -EINVAL;
-    }
-    if (period_nsec == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: create_thread called "
-			"with period of zero\n");
-	return -EINVAL;
-    }
-
-    if (strlen(name) > HAL_NAME_LEN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: thread name '%s' is too long\n", name);
-	return -EINVAL;
-    }
-    if (hal_data->lock & HAL_LOCK_CONFIG) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: create_thread called while"
-			" HAL is locked\n");
-	return -EPERM;
-    }
-    {
-	int  cmp  __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	/* make sure name is unique on thread list */
-	next = hal_data->thread_list_ptr;
-	while (next != 0) {
-	    tptr = SHMPTR(next);
-	    cmp = strcmp(tptr->name, name);
-	    if (cmp == 0) {
-		/* name already in list, can't insert */
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: duplicate thread name %s\n", name);
-		return -EINVAL;
-	    }
-	    /* didn't find it yet, look at next one */
-	    next = tptr->next_ptr;
-	}
-	/* allocate a new thread structure */
-	new = alloc_thread_struct();
-	if (new == 0) {
-	    /* alloc failed */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory to create thread\n");
-	    return -ENOMEM;
-	}
-	/* initialize the structure */
-	new->uses_fp = uses_fp;
-	new->cpu_id = cpu_id;
-	new->handle = rtapi_next_handle();
-	rtapi_snprintf(new->name, sizeof(new->name), "%s", name);
-	/* have to create and start a task to run the thread */
-	if (hal_data->thread_list_ptr == 0) {
-	    /* this is the first thread created */
-	    /* is timer started? if so, what period? */
-	    curr_period = rtapi_clock_set_period(0);
-	    if (curr_period == 0) {
-		/* not running, start it */
-		curr_period = rtapi_clock_set_period(period_nsec);
-		if (curr_period < 0) {
-		    hal_print_msg(RTAPI_MSG_ERR,
-				    "HAL_LIB: ERROR: clock_set_period returned %ld\n",
-				    curr_period);
-		    return -EINVAL;
-		}
-	    }
-	    /* make sure period <= desired period (allow 1% roundoff error) */
-	    if (curr_period > (period_nsec + (period_nsec / 100))) {
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL_LIB: ERROR: clock period too long: %ld\n", curr_period);
-		return -EINVAL;
-	    }
-	    if(hal_data->exact_base_period) {
-		hal_data->base_period = period_nsec;
-	    } else {
-		hal_data->base_period = curr_period;
-	    }
-	    /* reserve the highest priority (maybe for a watchdog?) */
-	    prev_priority = rtapi_prio_highest();
-	    /* no previous period to worry about */
-	    prev_period = 0;
-	} else {
-	    /* there are other threads, slowest (and lowest
-	       priority) is at head of list */
-	    tptr = SHMPTR(hal_data->thread_list_ptr);
-	    prev_period = tptr->period;
-	    prev_priority = tptr->priority;
-	}
-	if ( period_nsec < hal_data->base_period) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL_LIB: ERROR: new thread period %ld is less than clock period %ld\n",
-			    period_nsec, hal_data->base_period);
-	    return -EINVAL;
-	}
-	/* make period an integer multiple of the timer period */
-	n = (period_nsec + hal_data->base_period / 2) / hal_data->base_period;
-	new->period = hal_data->base_period * n;
-	if ( new->period < prev_period ) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL_LIB: ERROR: new thread period %ld is less than existing thread period %ld\n",
-			    period_nsec, prev_period);
-	    return -EINVAL;
-	}
-	/* make priority one lower than previous */
-	new->priority = rtapi_prio_next_lower(prev_priority);
-	/* create task - owned by library module, not caller */
-	retval = rtapi_task_new(thread_task,
-				new,
-				new->priority,
-				lib_module_id,
-				global_data->hal_thread_stack_size,
-				uses_fp,
-				new->name, new->cpu_id);
-	if (retval < 0) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL_LIB: could not create task for thread %s\n", name);
-	    return -EINVAL;
-	}
-	new->task_id = retval;
-	/* start task */
-	retval = rtapi_task_start(new->task_id, new->period);
-	if (retval < 0) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL_LIB: could not start task for thread %s: %d\n", name, retval);
-	    return -EINVAL;
-	}
-	/* insert new structure at head of list */
-	new->next_ptr = hal_data->thread_list_ptr;
-	hal_data->thread_list_ptr = SHMOFF(new);
-
-	// exit block protected by scoped lock
-    }
-
-    /* init time logging variables */
-    new->runtime = 0;
-    new->maxtime = 0;
-    /*! \todo Another #if 0 */
-#if 0
-    /* These params need to be re-visited when I refactor HAL.  Right
-       now they cause problems - they can no longer be owned by the calling
-       component, and they can't be owned by the hal_lib because it isn't
-       actually a component.
-    */
-    /* create a parameter with the thread's runtime in it */
-    rtapi_snprintf(buf, sizeof(buf), "%s.time", name);
-    hal_param_s32_new(buf, HAL_RO, &(new->runtime), lib_module_id);
-    /* create a parameter with the thread's maximum runtime in it */
-    rtapi_snprintf(buf, sizeof(buf), "%s.tmax", name);
-    hal_param_s32_new(buf, HAL_RW, &(new->maxtime), lib_module_id);
-#endif
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: thread %s id %d created prio=%d\n",
-		    name, new->task_id, new->priority);
-    return 0;
-}
-
-extern int hal_thread_delete(const char *name)
-{
-    int *prev, next;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: thread_delete called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_CONFIG) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: thread_delete called while HAL is locked\n");
-	return -EPERM;
-    }
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: deleting thread '%s'\n", name);
-    {
-	hal_thread_t *thread __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing shared data */
-	rtapi_mutex_get(&(hal_data->mutex));
-	/* search for the signal */
-	prev = &(hal_data->thread_list_ptr);
-	next = *prev;
-	while (next != 0) {
-	    thread = SHMPTR(next);
-	    if (strcmp(thread->name, name) == 0) {
-		/* this is the right thread, unlink from list */
-		*prev = thread->next_ptr;
-		/* and delete it */
-		free_thread_struct(thread);
-		/* done */
-		return 0;
-	    }
-	    /* no match, try the next one */
-	    prev = &(thread->next_ptr);
-	    next = *prev;
-	}
-    }
-    /* if we get here, we didn't find a match */
-    hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: thread '%s' not found\n",
-		    name);
-    return -EINVAL;
-}
-
-#endif /* RTAPI */
-
-int hal_add_funct_to_thread(const char *funct_name,
-			    const char *thread_name, int position)
-{
-    hal_funct_t *funct;
-    hal_list_t *list_root, *list_entry;
-    int n;
-    hal_funct_entry_t *funct_entry;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: add_funct called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_CONFIG) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: add_funct_to_thread called"
-			" while HAL is locked\n");
-	return -EPERM;
-    }
-
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: adding function '%s' to thread '%s'\n",
-		    funct_name, thread_name);
-    {
-	hal_thread_t *thread __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing data structures */
-	rtapi_mutex_get(&(hal_data->mutex));
-
-	/* make sure position is valid */
-	if (position == 0) {
-	    /* zero is not allowed */
-	    hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: bad position: 0\n");
-	    return -EINVAL;
-	}
-	/* make sure we were given a function name */
-	if (funct_name == 0) {
-	    /* no name supplied */
-	    hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: missing function name\n");
-	    return -EINVAL;
-	}
-	/* make sure we were given a thread name */
-	if (thread_name == 0) {
-	    /* no name supplied */
-	    hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: missing thread name\n");
-	    return -EINVAL;
-	}
-	/* search function list for the function */
-	funct = halpr_find_funct_by_name(funct_name);
-	if (funct == 0) {
-	    /* function not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: function '%s' not found\n",
-			    funct_name);
-	    return -EINVAL;
-	}
-	/* found the function, is it available? */
-	if ((funct->users > 0) && (funct->reentrant == 0)) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: function '%s' may only be added "
-			    "to one thread\n", funct_name);
-	    return -EINVAL;
-	}
-	/* search thread list for thread_name */
-	thread = halpr_find_thread_by_name(thread_name);
-	if (thread == 0) {
-	    /* thread not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: thread '%s' not found\n", thread_name);
-	    return -EINVAL;
-	}
-	/* ok, we have thread and function, are they compatible? */
-	if ((funct->uses_fp) && (!thread->uses_fp)) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: function '%s' needs FP\n", funct_name);
-	    return -EINVAL;
-	}
-	/* find insertion point */
-	list_root = &(thread->funct_list);
-	list_entry = list_root;
-	n = 0;
-	if (position > 0) {
-	    /* insertion is relative to start of list */
-	    while (++n < position) {
-		/* move further into list */
-		list_entry = list_next(list_entry);
-		if (list_entry == list_root) {
-		    /* reached end of list */
-		    hal_print_msg(RTAPI_MSG_ERR,
-				    "HAL: ERROR: position '%d' is too high\n", position);
-		    return -EINVAL;
-		}
-	    }
-	} else {
-	    /* insertion is relative to end of list */
-	    while (--n > position) {
-		/* move further into list */
-		list_entry = list_prev(list_entry);
-		if (list_entry == list_root) {
-		    /* reached end of list */
-		    hal_print_msg(RTAPI_MSG_ERR,
-				    "HAL: ERROR: position '%d' is too low\n", position);
-		    return -EINVAL;
-		}
-	    }
-	    /* want to insert before list_entry, so back up one more step */
-	    list_entry = list_prev(list_entry);
-	}
-	/* allocate a funct entry structure */
-	funct_entry = alloc_funct_entry_struct();
-	if (funct_entry == 0) {
-	    /* alloc failed */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: insufficient memory for thread->function link\n");
-	    return -ENOMEM;
-	}
-	/* init struct contents */
-	funct_entry->funct_ptr = SHMOFF(funct);
-	funct_entry->arg = funct->arg;
-	funct_entry->funct = funct->funct;
-	/* add the entry to the list */
-	list_add_after((hal_list_t *) funct_entry, list_entry);
-	/* update the function usage count */
-	funct->users++;
-    }
-    return 0;
-}
-
-int hal_del_funct_from_thread(const char *funct_name, const char *thread_name)
-{
-    hal_funct_t *funct;
-    hal_list_t *list_root, *list_entry;
-    hal_funct_entry_t *funct_entry;
-
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: del_funct called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_CONFIG) {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: ERROR: del_funct_from_thread called "
-			"while HAL is locked\n");
-	return -EPERM;
-    }
-
-    hal_print_msg(RTAPI_MSG_DBG,
-		    "HAL: removing function '%s' from thread '%s'\n",
-		    funct_name, thread_name);
-    {
-	hal_thread_t *thread __attribute__((cleanup(halpr_autorelease_mutex)));
-
-	/* get mutex before accessing data structures */
-	rtapi_mutex_get(&(hal_data->mutex));
-	/* make sure we were given a function name */
-	if (funct_name == 0) {
-	    /* no name supplied */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: missing function name\n");
-	    return -EINVAL;
-	}
-	/* make sure we were given a thread name */
-	if (thread_name == 0) {
-	    /* no name supplied */
-	    hal_print_msg(RTAPI_MSG_ERR, "HAL: ERROR: missing thread name\n");
-	    return -EINVAL;
-	}
-	/* search function list for the function */
-	funct = halpr_find_funct_by_name(funct_name);
-	if (funct == 0) {
-	    /* function not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: function '%s' not found\n", funct_name);
-	    return -EINVAL;
-	}
-	/* found the function, is it in use? */
-	if (funct->users == 0) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: function '%s' is not in use\n",
-			    funct_name);
-	    return -EINVAL;
-	}
-	/* search thread list for thread_name */
-	thread = halpr_find_thread_by_name(thread_name);
-	if (thread == 0) {
-	    /* thread not found */
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL: ERROR: thread '%s' not found\n", thread_name);
-	    return -EINVAL;
-	}
-	/* ok, we have thread and function, does thread use funct? */
-	list_root = &(thread->funct_list);
-	list_entry = list_next(list_root);
-	while (1) {
-	    if (list_entry == list_root) {
-		/* reached end of list, funct not found */
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL: ERROR: thread '%s' doesn't use %s\n",
-				thread_name, funct_name);
-		return -EINVAL;
-	    }
-	    funct_entry = (hal_funct_entry_t *) list_entry;
-	    if (SHMPTR(funct_entry->funct_ptr) == funct) {
-		/* this funct entry points to our funct, unlink */
-		list_remove_entry(list_entry);
-		/* and delete it */
-		free_funct_entry_struct(funct_entry);
-		/* done */
-		return 0;
-	    }
-	    /* try next one */
-	    list_entry = list_next(list_entry);
-	}
-    }
-}
-
-int hal_start_threads(void)
-{
-    /* a trivial function for a change! */
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: start_threads called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_RUN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: start_threads called while HAL is locked\n");
-	return -EPERM;
-    }
-
-
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: starting threads\n");
-    hal_data->threads_running = 1;
-    return 0;
-}
-
-int hal_stop_threads(void)
-{
-    /* wow, two in a row! */
-    if (hal_data == 0) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: stop_threads called before init\n");
-	return -EINVAL;
-    }
-
-    if (hal_data->lock & HAL_LOCK_RUN) {
-	hal_print_msg(RTAPI_MSG_ERR,
-	    "HAL: ERROR: stop_threads called while HAL is locked\n");
-	return -EPERM;
-    }
-
-    hal_data->threads_running = 0;
-    hal_print_msg(RTAPI_MSG_DBG, "HAL: threads stopped\n");
-    return 0;
-}
-
-/***********************************************************************
-*                    PRIVATE FUNCTION CODE                             *
-************************************************************************/
-
-hal_list_t *list_prev(hal_list_t * entry)
-{
-    /* this function is only needed because of memory mapping */
-    return SHMPTR(entry->prev);
-}
-
-hal_list_t *list_next(hal_list_t * entry)
-{
-    /* this function is only needed because of memory mapping */
-    return SHMPTR(entry->next);
-}
-
-void list_init_entry(hal_list_t * entry)
-{
-    int entry_n;
-
-    entry_n = SHMOFF(entry);
-    entry->next = entry_n;
-    entry->prev = entry_n;
-}
-
-void list_add_after(hal_list_t * entry, hal_list_t * prev)
-{
-    int entry_n, prev_n, next_n;
-    hal_list_t *next;
-
-    /* messiness needed because of memory mapping */
-    entry_n = SHMOFF(entry);
-    prev_n = SHMOFF(prev);
-    next_n = prev->next;
-    next = SHMPTR(next_n);
-    /* insert the entry */
-    entry->next = next_n;
-    entry->prev = prev_n;
-    prev->next = entry_n;
-    next->prev = entry_n;
-}
-
-void list_add_before(hal_list_t * entry, hal_list_t * next)
-{
-    int entry_n, prev_n, next_n;
-    hal_list_t *prev;
-
-    /* messiness needed because of memory mapping */
-    entry_n = SHMOFF(entry);
-    next_n = SHMOFF(next);
-    prev_n = next->prev;
-    prev = SHMPTR(prev_n);
-    /* insert the entry */
-    entry->next = next_n;
-    entry->prev = prev_n;
-    prev->next = entry_n;
-    next->prev = entry_n;
-}
-
-hal_list_t *list_remove_entry(hal_list_t * entry)
-{
-    int entry_n;
-    hal_list_t *prev, *next;
-
-    /* messiness needed because of memory mapping */
-    entry_n = SHMOFF(entry);
-    prev = SHMPTR(entry->prev);
-    next = SHMPTR(entry->next);
-    /* remove the entry */
-    prev->next = entry->next;
-    next->prev = entry->prev;
-    entry->next = entry_n;
-    entry->prev = entry_n;
-    return next;
-}
-
-hal_comp_t *halpr_find_comp_by_name(const char *name)
-{
-    int next;
-    hal_comp_t *comp;
-
-    /* search component list for 'name' */
-    next = hal_data->comp_list_ptr;
-    while (next != 0) {
-	comp = SHMPTR(next);
-	if (strcmp(comp->name, name) == 0) {
-	    /* found a match */
-	    return comp;
-	}
-	/* didn't find it yet, look at next one */
-	next = comp->next_ptr;
-    }
-    /* if loop terminates, we reached end of list with no match */
-    return 0;
-}
-
-hal_pin_t *halpr_find_pin_by_name(const char *name)
-{
-    int next;
-    hal_pin_t *pin;
-    hal_oldname_t *oldname;
-
-    /* search pin list for 'name' */
-    next = hal_data->pin_list_ptr;
-    while (next != 0) {
-	pin = SHMPTR(next);
-	if (strcmp(pin->name, name) == 0) {
-	    /* found a match */
-	    return pin;
-	}
-	if (pin->oldname != 0 ) {
-	    oldname = SHMPTR(pin->oldname);
-	    if (strcmp(oldname->name, name) == 0) {
-		/* found a match */
-		return pin;
-	    }
-	}
-	/* didn't find it yet, look at next one */
-	next = pin->next_ptr;
-    }
-    /* if loop terminates, we reached end of list with no match */
-    return 0;
-}
-
-hal_sig_t *halpr_find_sig_by_name(const char *name)
-{
-    int next;
-    hal_sig_t *sig;
-
-    /* search signal list for 'name' */
-    next = hal_data->sig_list_ptr;
-    while (next != 0) {
-	sig = SHMPTR(next);
-	if (strcmp(sig->name, name) == 0) {
-	    /* found a match */
-	    return sig;
-	}
-	/* didn't find it yet, look at next one */
-	next = sig->next_ptr;
-    }
-    /* if loop terminates, we reached end of list with no match */
-    return 0;
-}
-
-hal_param_t *halpr_find_param_by_name(const char *name)
-{
-    int next;
-    hal_param_t *param;
-    hal_oldname_t *oldname;
-
-    /* search parameter list for 'name' */
-    next = hal_data->param_list_ptr;
-    while (next != 0) {
-	param = SHMPTR(next);
-	if (strcmp(param->name, name) == 0) {
-	    /* found a match */
-	    return param;
-	}
-	if (param->oldname != 0 ) {
-	    oldname = SHMPTR(param->oldname);
-	    if (strcmp(oldname->name, name) == 0) {
-		/* found a match */
-		return param;
-	    }
-	}
-	/* didn't find it yet, look at next one */
-	next = param->next_ptr;
-    }
-    /* if loop terminates, we reached end of list with no match */
-    return 0;
-}
-
-hal_thread_t *halpr_find_thread_by_name(const char *name)
-{
-    int next;
-    hal_thread_t *thread;
-
-    /* search thread list for 'name' */
-    next = hal_data->thread_list_ptr;
-    while (next != 0) {
-	thread = SHMPTR(next);
-	if (strcmp(thread->name, name) == 0) {
-	    /* found a match */
-	    return thread;
-	}
-	/* didn't find it yet, look at next one */
-	next = thread->next_ptr;
-    }
-    /* if loop terminates, we reached end of list with no match */
-    return 0;
-}
-
-hal_funct_t *halpr_find_funct_by_name(const char *name)
-{
-    int next;
-    hal_funct_t *funct;
-
-    /* search function list for 'name' */
-    next = hal_data->funct_list_ptr;
-    while (next != 0) {
-	funct = SHMPTR(next);
-	if (strcmp(funct->name, name) == 0) {
-	    /* found a match */
-	    return funct;
-	}
-	/* didn't find it yet, look at next one */
-	next = funct->next_ptr;
-    }
-    /* if loop terminates, we reached end of list with no match */
-    return 0;
-}
-
-hal_comp_t *halpr_find_comp_by_id(int id)
-{
-    int next;
-    hal_comp_t *comp;
-
-    /* search list for 'comp_id' */
-    next = hal_data->comp_list_ptr;
-    while (next != 0) {
-	comp = SHMPTR(next);
-	if (comp->comp_id == id) {
-	    /* found a match */
-	    return comp;
-	}
-	/* didn't find it yet, look at next one */
-	next = comp->next_ptr;
-    }
-    /* if loop terminates, we reached end of list without finding a match */
-    return 0;
-}
-
-hal_pin_t *halpr_find_pin_by_owner(hal_comp_t * owner, hal_pin_t * start)
-{
-    int owner_ptr, next;
-    hal_pin_t *pin;
-
-    /* get offset of 'owner' component */
-    owner_ptr = SHMOFF(owner);
-    /* is this the first call? */
-    if (start == 0) {
-	/* yes, start at beginning of pin list */
-	next = hal_data->pin_list_ptr;
-    } else {
-	/* no, start at next pin */
-	next = start->next_ptr;
-    }
-    while (next != 0) {
-	pin = SHMPTR(next);
-	if (pin->owner_ptr == owner_ptr) {
-	    /* found a match */
-	    return pin;
-	}
-	/* didn't find it yet, look at next one */
-	next = pin->next_ptr;
-    }
-    /* if loop terminates, we reached end of list without finding a match */
-    return 0;
-}
-
-hal_param_t *halpr_find_param_by_owner(hal_comp_t * owner,
-    hal_param_t * start)
-{
-    int owner_ptr, next;
-    hal_param_t *param;
-
-    /* get offset of 'owner' component */
-    owner_ptr = SHMOFF(owner);
-    /* is this the first call? */
-    if (start == 0) {
-	/* yes, start at beginning of param list */
-	next = hal_data->param_list_ptr;
-    } else {
-	/* no, start at next param */
-	next = start->next_ptr;
-    }
-    while (next != 0) {
-	param = SHMPTR(next);
-	if (param->owner_ptr == owner_ptr) {
-	    /* found a match */
-	    return param;
-	}
-	/* didn't find it yet, look at next one */
-	next = param->next_ptr;
-    }
-    /* if loop terminates, we reached end of list without finding a match */
-    return 0;
-}
-
-hal_funct_t *halpr_find_funct_by_owner(hal_comp_t * owner,
-    hal_funct_t * start)
-{
-    int owner_ptr, next;
-    hal_funct_t *funct;
-
-    /* get offset of 'owner' component */
-    owner_ptr = SHMOFF(owner);
-    /* is this the first call? */
-    if (start == 0) {
-	/* yes, start at beginning of function list */
-	next = hal_data->funct_list_ptr;
-    } else {
-	/* no, start at next function */
-	next = start->next_ptr;
-    }
-    while (next != 0) {
-	funct = SHMPTR(next);
-	if (funct->owner_ptr == owner_ptr) {
-	    /* found a match */
-	    return funct;
-	}
-	/* didn't find it yet, look at next one */
-	next = funct->next_ptr;
-    }
-    /* if loop terminates, we reached end of list without finding a match */
-    return 0;
-}
-
-hal_pin_t *halpr_find_pin_by_sig(hal_sig_t * sig, hal_pin_t * start)
-{
-    int sig_ptr, next;
-    hal_pin_t *pin;
-
-    /* get offset of 'sig' component */
-    sig_ptr = SHMOFF(sig);
-    /* is this the first call? */
-    if (start == 0) {
-	/* yes, start at beginning of pin list */
-	next = hal_data->pin_list_ptr;
-    } else {
-	/* no, start at next pin */
-	next = start->next_ptr;
-    }
-    while (next != 0) {
-	pin = SHMPTR(next);
-	if (pin->signal == sig_ptr) {
-	    /* found a match */
-	    return pin;
-	}
-	/* didn't find it yet, look at next one */
-	next = pin->next_ptr;
-    }
-    /* if loop terminates, we reached end of list without finding a match */
-    return 0;
-}
-
-
 /***********************************************************************
 *         Scope exit unlock helper                                     *
 *         see hal_priv.h for usage hints                               *
@@ -2573,124 +196,6 @@ void halpr_autorelease_mutex(void *variable)
 /* these functions are called when the hal_lib module is insmod'ed
    or rmmod'ed.
 */
-#if defined(BUILD_SYS_USER_DSO)
-#undef CONFIG_PROC_FS
-#endif
-
-#ifdef CONFIG_PROC_FS
-#include <linux/proc_fs.h>
-extern struct proc_dir_entry *rtapi_dir;
-static struct proc_dir_entry *hal_dir = 0;
-static struct proc_dir_entry *hal_newinst_file = 0;
-static struct proc_dir_entry *hal_threadcmd = 0;
-
-static int proc_write_newinst(struct file *file,
-        const char *buffer, unsigned long count, void *data)
-{
-    if(hal_data->pending_constructor) {
-        hal_print_msg(RTAPI_MSG_DBG,
-                "HAL: running constructor for %s %s\n",
-                hal_data->constructor_prefix,
-                hal_data->constructor_arg);
-        hal_data->pending_constructor(hal_data->constructor_prefix,
-                hal_data->constructor_arg);
-        hal_data->pending_constructor = 0;
-    }
-    return count;
-}
-
-// simple interface to hal_create_thread()/hal_thread_delete()
-// through /proc/rtapi/hal/threadcmd (kernel threadstyles only)
-//
-// to start a thread, write 'newthread' <threadname> <period> <fp> <cpu>'
-// example:
-//    echo newthread servo-thread 1000000 1 -1 >/proc/rtapi/hal/threadcmd
-//
-// to delete a thread, write 'delthread <threadname>'
-//    echo delthread servo-thread >/proc/rtapi/hal/threadcmd
-//
-// HAL return values are reflected in the return value to write()
-//
-static int proc_write_threadcmd(struct file *file,
-        const char *buffer, unsigned long count, void *data)
-{
-    char cmd[20], name[HAL_NAME_LEN + 1];
-    unsigned long period;
-    int fp, cpu, retval;
-
-    if (!strncmp(buffer,"newthread", 9)) {
-	if ((retval = sscanf(buffer, "%s %s %lu %d %d",
-			     cmd, name, &period, &fp, &cpu)) != 5) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL:newthread: expecting 5 items (s:cmd s:name d:period d:fp d:cpu), got %d\n",
-			    retval);
-	    return -EINVAL;
-	}
-	if ((period > 0) &&
-	    (strlen(name) > 0)) {
-	    retval = hal_create_thread(name, period, fp, cpu);
-	    if (retval < 0) {
-		hal_print_msg(RTAPI_MSG_ERR,
-				"HAL:newthread: could not create thread '%s' - error %d\n",
-				name, retval);
-		return retval;
-	    } else {
-		hal_print_msg(RTAPI_MSG_INFO,
-				"HAL:newthread: created %ld uS thread '%s' fp=%d cpu=%d\n",
-				period / 1000, name, fp, cpu);
-	    }
-	}
-    } else if (!strncmp(buffer, "delthread", 9)) {
-	if ((retval = sscanf(buffer, "%s %s", cmd, name)) != 2) {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL:delthread: expecting 2 items: 'delthread <threadname>'\n");
-	    return -EINVAL;
-	}
-	if ((retval = hal_thread_delete(name)))  {
-	    hal_print_msg(RTAPI_MSG_ERR,
-			    "HAL:delthread '%s' error %d\n", name, retval);
-	    return retval;
-	}
-	hal_print_msg(RTAPI_MSG_INFO,
-			"HAL:delthread - thread '%s' deleted\n", name);
-    } else {
-	hal_print_msg(RTAPI_MSG_ERR,
-			"HAL: unrecognized threadcmd: '%s'\n", cmd);
-	return -EINVAL;
-    }
-    return count;
-}
-
-static void hal_proc_clean(void) {
-    if(hal_newinst_file)
-        remove_proc_entry("newinst", hal_dir);
-    if(hal_threadcmd)
-        remove_proc_entry("threadcmd", hal_dir);
-    if(hal_dir)
-        remove_proc_entry("hal", rtapi_dir);
-    hal_newinst_file = hal_dir = hal_threadcmd = 0;
-}
-
-static int hal_proc_init(void) {
-    if(!rtapi_dir) return 0;
-    hal_dir = create_proc_entry("hal", S_IFDIR, rtapi_dir);
-    if(!hal_dir) { hal_proc_clean(); return -1; }
-    hal_newinst_file = create_proc_entry("newinst", 0666, hal_dir);
-    if(!hal_newinst_file) { hal_proc_clean(); return -1; }
-    hal_newinst_file->data = NULL;
-    hal_newinst_file->read_proc = NULL;
-    hal_newinst_file->write_proc = proc_write_newinst;
-    hal_threadcmd = create_proc_entry("threadcmd", 0666, hal_dir);
-    if(!hal_threadcmd) { hal_proc_clean(); return -1; }
-    hal_threadcmd->data = NULL;
-    hal_threadcmd->read_proc = NULL;
-    hal_threadcmd->write_proc = proc_write_threadcmd;
-    return 0;
-}
-#else
-static int hal_proc_clean(void) { return 0; }
-static int hal_proc_init(void) { return 0; }
-#endif
 
 
 int rtapi_app_main(void)
@@ -2762,7 +267,6 @@ int rtapi_app_main(void)
 	rtapi_exit(lib_module_id);
 	return -EINVAL;
     }
-    //rtapi_printall();
 
     /* done */
     hal_print_msg(RTAPI_MSG_DBG,
@@ -2815,57 +319,6 @@ void rtapi_app_exit(void)
 		    rtapi_instance);
 }
 
-/* this is the task function that implements threads in realtime */
-
-static void thread_task(void *arg)
-{
-    hal_thread_t *thread;
-    hal_funct_t *funct;
-    hal_funct_entry_t *funct_root, *funct_entry;
-    long long int start_time, end_time;
-    long long int thread_start_time;
-
-    thread = arg;
-    while (1) {
-	if (hal_data->threads_running > 0) {
-	    /* point at first function on function list */
-	    funct_root = (hal_funct_entry_t *) & (thread->funct_list);
-	    funct_entry = SHMPTR(funct_root->links.next);
-	    /* execution time logging */
-	    start_time = rtapi_get_clocks();
-	    end_time = start_time;
-	    thread_start_time = start_time;
-	    /* run thru function list */
-	    while (funct_entry != funct_root) {
-		/* call the function */
-		funct_entry->funct(funct_entry->arg, thread->period);
-		/* capture execution time */
-		end_time = rtapi_get_clocks();
-		/* point to function structure */
-		funct = SHMPTR(funct_entry->funct_ptr);
-		/* update execution time data */
-		*(funct->runtime) = (hal_s32_t)(end_time - start_time);
-		if ( *(funct->runtime) > funct->maxtime) {
-		    funct->maxtime = *(funct->runtime);
-		    funct->maxtime_increased = 1;
-		} else {
-		    funct->maxtime_increased = 0;
-		}
-		/* point to next next entry in list */
-		funct_entry = SHMPTR(funct_entry->links.next);
-		/* prepare to measure time for next funct */
-		start_time = end_time;
-	    }
-	    /* update thread execution time */
-	    thread->runtime = (hal_s32_t)(end_time - thread_start_time);
-	    if (thread->runtime > thread->maxtime) {
-		thread->maxtime = thread->runtime;
-	    }
-	}
-	/* wait until next period */
-	rtapi_wait();
-    }
-}
 #endif /* RTAPI */
 
 /* see the declarations of these functions (near top of file) for
@@ -2944,690 +397,6 @@ static int init_hal_data(void)
     /* done, release mutex */
     rtapi_mutex_give(&(hal_data->mutex));
     return 0;
-}
-
-void *shmalloc_up(long int size)
-{
-    long int tmp_bot;
-    void *retval;
-
-    /* deal with alignment requirements */
-    tmp_bot = hal_data->shmem_bot;
-    if (size >= 8) {
-	/* align on 8 byte boundary */
-	tmp_bot = (tmp_bot + 7) & (~7);
-    } else if (size >= 4) {
-	/* align on 4 byte boundary */
-	tmp_bot = (tmp_bot + 3) & (~3);
-    } else if (size == 2) {
-	/* align on 2 byte boundary */
-	tmp_bot = (tmp_bot + 1) & (~1);
-    }
-    /* is there enough memory available? */
-    if ((hal_data->shmem_top - tmp_bot) < size) {
-	/* no */
-	return 0;
-    }
-    /* memory is available, allocate it */
-    retval = SHMPTR(tmp_bot);
-    hal_data->shmem_bot = tmp_bot + size;
-    hal_data->shmem_avail = hal_data->shmem_top - hal_data->shmem_bot;
-    return retval;
-}
-
-void *shmalloc_dn(long int size)
-{
-    long int tmp_top;
-    void *retval;
-
-    /* tentatively allocate memory */
-    tmp_top = hal_data->shmem_top - size;
-    /* deal with alignment requirements */
-    if (size >= 8) {
-	/* align on 8 byte boundary */
-	tmp_top &= (~7);
-    } else if (size >= 4) {
-	/* align on 4 byte boundary */
-	tmp_top &= (~3);
-    } else if (size == 2) {
-	/* align on 2 byte boundary */
-	tmp_top &= (~1);
-    }
-    /* is there enough memory available? */
-    if (tmp_top < hal_data->shmem_bot) {
-	/* no */
-	return 0;
-    }
-    /* memory is available, allocate it */
-    retval = SHMPTR(tmp_top);
-    hal_data->shmem_top = tmp_top;
-    hal_data->shmem_avail = hal_data->shmem_top - hal_data->shmem_bot;
-    return retval;
-}
-
-hal_comp_t *halpr_alloc_comp_struct(void)
-{
-    hal_comp_t *p;
-
-    /* check the free list */
-    if (hal_data->comp_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->comp_free_ptr);
-	/* unlink it from the free list */
-	hal_data->comp_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_comp_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->comp_id = 0;
-	p->type = TYPE_INVALID;
-	p->state = COMP_INVALID;
-	p->shmem_base = 0;
-	p->name[0] = '\0';
-    }
-    return p;
-}
-
-static hal_pin_t *alloc_pin_struct(void)
-{
-    hal_pin_t *p;
-
-    /* check the free list */
-    if (hal_data->pin_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->pin_free_ptr);
-	/* unlink it from the free list */
-	hal_data->pin_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_pin_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->data_ptr_addr = 0;
-	p->owner_ptr = 0;
-	p->type = 0;
-	p->dir = 0;
-	p->signal = 0;
-	memset(&p->dummysig, 0, sizeof(hal_data_u));
-	p->name[0] = '\0';
-	p->eps_index = 0;
-	p->flags = 0;
-    }
-    return p;
-}
-
-static hal_sig_t *alloc_sig_struct(void)
-{
-    hal_sig_t *p;
-
-    /* check the free list */
-    if (hal_data->sig_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->sig_free_ptr);
-	/* unlink it from the free list */
-	hal_data->sig_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_sig_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->data_ptr = 0;
-	p->type = 0;
-	p->readers = 0;
-	p->writers = 0;
-	p->bidirs = 0;
-	p->name[0] = '\0';
-    }
-    return p;
-}
-
-static hal_param_t *alloc_param_struct(void)
-{
-    hal_param_t *p;
-
-    /* check the free list */
-    if (hal_data->param_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->param_free_ptr);
-	/* unlink it from the free list */
-	hal_data->param_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_param_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->data_ptr = 0;
-	p->owner_ptr = 0;
-	p->type = 0;
-	p->name[0] = '\0';
-    }
-    return p;
-}
-
-static hal_oldname_t *halpr_alloc_oldname_struct(void)
-{
-    hal_oldname_t *p;
-
-    /* check the free list */
-    if (hal_data->oldname_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->oldname_free_ptr);
-	/* unlink it from the free list */
-	hal_data->oldname_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_oldname_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->name[0] = '\0';
-    }
-    return p;
-}
-
-#ifdef RTAPI
-static hal_funct_t *alloc_funct_struct(void)
-{
-    hal_funct_t *p;
-
-    /* check the free list */
-    if (hal_data->funct_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->funct_free_ptr);
-	/* unlink it from the free list */
-	hal_data->funct_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_funct_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->uses_fp = 0;
-	p->owner_ptr = 0;
-	p->reentrant = 0;
-	p->users = 0;
-	p->arg = 0;
-	p->funct = 0;
-	p->name[0] = '\0';
-    }
-    return p;
-}
-#endif /* RTAPI */
-
-static hal_funct_entry_t *alloc_funct_entry_struct(void)
-{
-    hal_list_t *freelist, *l;
-    hal_funct_entry_t *p;
-
-    /* check the free list */
-    freelist = &(hal_data->funct_entry_free);
-    l = list_next(freelist);
-    if (l != freelist) {
-	/* found a free structure, unlink from the free list */
-	list_remove_entry(l);
-	p = (hal_funct_entry_t *) l;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_funct_entry_t));
-	l = (hal_list_t *) p;
-	list_init_entry(l);
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->funct_ptr = 0;
-	p->arg = 0;
-	p->funct = 0;
-    }
-    return p;
-}
-
-#ifdef RTAPI
-static hal_thread_t *alloc_thread_struct(void)
-{
-    hal_thread_t *p;
-
-    /* check the free list */
-    if (hal_data->thread_free_ptr != 0) {
-	/* found a free structure, point to it */
-	p = SHMPTR(hal_data->thread_free_ptr);
-	/* unlink it from the free list */
-	hal_data->thread_free_ptr = p->next_ptr;
-	p->next_ptr = 0;
-    } else {
-	/* nothing on free list, allocate a brand new one */
-	p = shmalloc_dn(sizeof(hal_thread_t));
-    }
-    if (p) {
-	/* make sure it's empty */
-	p->next_ptr = 0;
-	p->uses_fp = 0;
-	p->period = 0;
-	p->priority = 0;
-	p->task_id = 0;
-	list_init_entry(&(p->funct_list));
-	p->name[0] = '\0';
-    }
-    return p;
-}
-#endif /* RTAPI */
-
-static void free_comp_struct(hal_comp_t * comp)
-{
-    int *prev, next;
-#ifdef RTAPI
-    hal_funct_t *funct;
-#endif /* RTAPI */
-    hal_pin_t *pin;
-    hal_param_t *param;
-
-    /* can't delete the component until we delete its "stuff" */
-    /* need to check for functs only if a realtime component */
-#ifdef RTAPI
-    /* search the function list for this component's functs */
-    prev = &(hal_data->funct_list_ptr);
-    next = *prev;
-    while (next != 0) {
-	funct = SHMPTR(next);
-	if (SHMPTR(funct->owner_ptr) == comp) {
-	    /* this function belongs to our component, unlink from list */
-	    *prev = funct->next_ptr;
-	    /* and delete it */
-	    free_funct_struct(funct);
-	} else {
-	    /* no match, try the next one */
-	    prev = &(funct->next_ptr);
-	}
-	next = *prev;
-    }
-#endif /* RTAPI */
-    /* search the pin list for this component's pins */
-    prev = &(hal_data->pin_list_ptr);
-    next = *prev;
-    while (next != 0) {
-	pin = SHMPTR(next);
-	if (SHMPTR(pin->owner_ptr) == comp) {
-	    /* this pin belongs to our component, unlink from list */
-	    *prev = pin->next_ptr;
-	    /* and delete it */
-	    free_pin_struct(pin);
-	} else {
-	    /* no match, try the next one */
-	    prev = &(pin->next_ptr);
-	}
-	next = *prev;
-    }
-    /* search the parameter list for this component's parameters */
-    prev = &(hal_data->param_list_ptr);
-    next = *prev;
-    while (next != 0) {
-	param = SHMPTR(next);
-	if (SHMPTR(param->owner_ptr) == comp) {
-	    /* this param belongs to our component, unlink from list */
-	    *prev = param->next_ptr;
-	    /* and delete it */
-	    free_param_struct(param);
-	} else {
-	    /* no match, try the next one */
-	    prev = &(param->next_ptr);
-	}
-	next = *prev;
-    }
-    /* now we can delete the component itself */
-    /* clear contents of struct */
-    comp->comp_id = -1;
-    comp->type = TYPE_INVALID;
-    comp->state = COMP_INVALID;
-    comp->last_bound = 0;
-    comp->last_unbound = 0;
-    comp->last_update = 0;
-    comp->shmem_base = 0;
-    comp->name[0] = '\0';
-    /* add it to free list */
-    comp->next_ptr = hal_data->comp_free_ptr;
-    hal_data->comp_free_ptr = SHMOFF(comp);
-}
-
-static void unlink_pin(hal_pin_t * pin)
-{
-    hal_sig_t *sig;
-    hal_comp_t *comp;
-    void *dummy_addr, **data_ptr_addr;
-    hal_data_u *sig_data_addr;
-
-    /* is this pin linked to a signal? */
-    if (pin->signal != 0) {
-	/* yes, need to unlink it */
-	sig = SHMPTR(pin->signal);
-	/* make pin's 'data_ptr' point to its dummy signal */
-	data_ptr_addr = SHMPTR(pin->data_ptr_addr);
-	comp = SHMPTR(pin->owner_ptr);
-	dummy_addr = comp->shmem_base + SHMOFF(&(pin->dummysig));
-	*data_ptr_addr = dummy_addr;
-
-	/* copy current signal value to dummy */
-	sig_data_addr = (hal_data_u *)(hal_shmem_base + sig->data_ptr);
-	dummy_addr = hal_shmem_base + SHMOFF(&(pin->dummysig));
-	*(hal_data_u *)dummy_addr = *sig_data_addr;
-
-	/* update the signal's reader/writer counts */
-	if ((pin->dir & HAL_IN) != 0) {
-	    sig->readers--;
-	}
-	if (pin->dir == HAL_OUT) {
-	    sig->writers--;
-	}
-	if (pin->dir == HAL_IO) {
-	    sig->bidirs--;
-	}
-	/* mark pin as unlinked */
-	pin->signal = 0;
-    }
-}
-
-static void free_pin_struct(hal_pin_t * pin)
-{
-
-    unlink_pin(pin);
-    /* clear contents of struct */
-    if ( pin->oldname != 0 ) free_oldname_struct(SHMPTR(pin->oldname));
-    pin->data_ptr_addr = 0;
-    pin->owner_ptr = 0;
-    pin->type = 0;
-    pin->dir = 0;
-    pin->signal = 0;
-    pin->handle = -1;
-    memset(&pin->dummysig, 0, sizeof(hal_data_u));
-    pin->name[0] = '\0';
-    pin->eps_index = 0;
-    pin->flags = 0;
-    /* add it to free list */
-    pin->next_ptr = hal_data->pin_free_ptr;
-    hal_data->pin_free_ptr = SHMOFF(pin);
-}
-
-static void free_sig_struct(hal_sig_t * sig)
-{
-    hal_pin_t *pin;
-
-    /* look for pins linked to this signal */
-    pin = halpr_find_pin_by_sig(sig, 0);
-    while (pin != 0) {
-	/* found one, unlink it */
-	unlink_pin(pin);
-	/* check for another pin linked to the signal */
-	pin = halpr_find_pin_by_sig(sig, pin);
-    }
-    /* clear contents of struct */
-    sig->data_ptr = 0;
-    sig->type = 0;
-    sig->readers = 0;
-    sig->writers = 0;
-    sig->bidirs = 0;
-    sig->handle = -1;
-    sig->name[0] = '\0';
-    /* add it to free list */
-    sig->next_ptr = hal_data->sig_free_ptr;
-    hal_data->sig_free_ptr = SHMOFF(sig);
-}
-
-static void free_param_struct(hal_param_t * p)
-{
-    /* clear contents of struct */
-    if ( p->oldname != 0 ) free_oldname_struct(SHMPTR(p->oldname));
-    p->data_ptr = 0;
-    p->owner_ptr = 0;
-    p->type = 0;
-    p->name[0] = '\0';
-    p->handle = -1;
-    /* add it to free list (params use the same struct as src vars) */
-    p->next_ptr = hal_data->param_free_ptr;
-    hal_data->param_free_ptr = SHMOFF(p);
-}
-
-static void free_oldname_struct(hal_oldname_t * oldname)
-{
-    /* clear contents of struct */
-    oldname->name[0] = '\0';
-    /* add it to free list */
-    oldname->next_ptr = hal_data->oldname_free_ptr;
-    hal_data->oldname_free_ptr = SHMOFF(oldname);
-}
-
-
-#ifdef RTAPI
-static void free_funct_struct(hal_funct_t * funct)
-{
-    int next_thread;
-    hal_thread_t *thread;
-    hal_list_t *list_root, *list_entry;
-    hal_funct_entry_t *funct_entry;
-
-/*  int next_thread, next_entry;*/
-
-    if (funct->users > 0) {
-	/* We can't casually delete the function, there are thread(s) which
-	   will call it.  So we must check all the threads and remove any
-	   funct_entrys that call this function */
-	/* start at root of thread list */
-	next_thread = hal_data->thread_list_ptr;
-	/* run through thread list */
-	while (next_thread != 0) {
-	    /* point to thread */
-	    thread = SHMPTR(next_thread);
-	    /* start at root of funct_entry list */
-	    list_root = &(thread->funct_list);
-	    list_entry = list_next(list_root);
-	    /* run thru funct_entry list */
-	    while (list_entry != list_root) {
-		/* point to funct entry */
-		funct_entry = (hal_funct_entry_t *) list_entry;
-		/* test it */
-		if (SHMPTR(funct_entry->funct_ptr) == funct) {
-		    /* this funct entry points to our funct, unlink */
-		    list_entry = list_remove_entry(list_entry);
-		    /* and delete it */
-		    free_funct_entry_struct(funct_entry);
-		} else {
-		    /* no match, try the next one */
-		    list_entry = list_next(list_entry);
-		}
-	    }
-	    /* move on to the next thread */
-	    next_thread = thread->next_ptr;
-	}
-    }
-    /* clear contents of struct */
-    funct->uses_fp = 0;
-    funct->owner_ptr = 0;
-    funct->reentrant = 0;
-    funct->users = 0;
-    funct->arg = 0;
-    funct->funct = 0;
-    funct->runtime = 0;
-    funct->name[0] = '\0';
-    /* add it to free list */
-    funct->next_ptr = hal_data->funct_free_ptr;
-    hal_data->funct_free_ptr = SHMOFF(funct);
-}
-#endif /* RTAPI */
-
-static void free_funct_entry_struct(hal_funct_entry_t * funct_entry)
-{
-    hal_funct_t *funct;
-
-    if (funct_entry->funct_ptr > 0) {
-	/* entry points to a function, update the function struct */
-	funct = SHMPTR(funct_entry->funct_ptr);
-	funct->users--;
-    }
-    /* clear contents of struct */
-    funct_entry->funct_ptr = 0;
-    funct_entry->arg = 0;
-    funct_entry->funct = 0;
-    /* add it to free list */
-    list_add_after((hal_list_t *) funct_entry, &(hal_data->funct_entry_free));
-}
-
-#ifdef RTAPI
-static void free_thread_struct(hal_thread_t * thread)
-{
-    hal_funct_entry_t *funct_entry;
-    hal_list_t *list_root, *list_entry;
-/*! \todo Another #if 0 */
-#if 0
-    int *prev, next;
-    char time[HAL_NAME_LEN + 1], tmax[HAL_NAME_LEN + 1];
-    hal_param_t *param;
-#endif
-
-    /* if we're deleting a thread, we need to stop all threads */
-    hal_data->threads_running = 0;
-    /* and stop the task associated with this thread */
-    rtapi_task_pause(thread->task_id);
-    rtapi_task_delete(thread->task_id);
-    /* clear contents of struct */
-    thread->uses_fp = 0;
-    thread->period = 0;
-    thread->priority = 0;
-    thread->task_id = 0;
-    /* clear the function entry list */
-    list_root = &(thread->funct_list);
-    list_entry = list_next(list_root);
-    while (list_entry != list_root) {
-	/* entry found, save pointer to it */
-	funct_entry = (hal_funct_entry_t *) list_entry;
-	/* unlink it, point to the next one */
-	list_entry = list_remove_entry(list_entry);
-	/* free the removed entry */
-	free_funct_entry_struct(funct_entry);
-    }
-/*! \todo Another #if 0 */
-#if 0
-/* Currently these don't get created, so we don't have to worry
-   about deleting them.  They will come back when the HAL refactor
-   is done, at that time this code or something like it will be
-   needed.
-*/
-    /* need to delete <thread>.time and <thread>.tmax params */
-    rtapi_snprintf(time, sizeof(time), "%s.time", thread->name);
-    rtapi_snprintf(tmax, sizeof(tmax), "%s.tmax", thread->name);
-    /* search the parameter list for those parameters */
-    prev = &(hal_data->param_list_ptr);
-    next = *prev;
-    while (next != 0) {
-	param = SHMPTR(next);
-	/* does this param match either name? */
-	if ((strcmp(param->name, time) == 0)
-	    || (strcmp(param->name, tmax) == 0)) {
-	    /* yes, unlink from list */
-	    *prev = param->next_ptr;
-	    /* and delete it */
-	    free_param_struct(param);
-	} else {
-	    /* no match, try the next one */
-	    prev = &(param->next_ptr);
-	}
-	next = *prev;
-    }
-#endif
-    thread->name[0] = '\0';
-    /* add thread to free list */
-    thread->next_ptr = hal_data->thread_free_ptr;
-    hal_data->thread_free_ptr = SHMOFF(thread);
-}
-#endif /* RTAPI */
-
-
-/***********************************************************************
-*                     utility functions, mostly used by haltalk        *
-*                                                                      *
-************************************************************************/
-
-
-// return number of pins in a component
-int
-halpr_pin_count(const char *name)
-{
-    hal_comp_t *comp;
-    hal_comp_t *owner;
-    hal_pin_t *pin;
-    int count = 0;
-
-    comp = halpr_find_comp_by_name(name);
-    if (comp == 0)
-	return -ENOENT;
-
-    int next = hal_data->pin_list_ptr;
-    while (next != 0) {
-	pin = (hal_pin_t *)SHMPTR(next);
-	owner = (hal_comp_t *) SHMPTR(pin->owner_ptr);
-	if (owner->comp_id == comp->comp_id)
-	    count++;
-	next = pin->next_ptr;
-    }
-    return count;
-}
-
-// return number of params in a component
-int
-halpr_param_count(const char *name)
-{
-    hal_comp_t *comp;
-    hal_comp_t *owner;
-    int count = 0;
-
-    comp = halpr_find_comp_by_name(name);
-    if (comp == 0)
-	return -ENOENT;
-
-    int next = hal_data->param_list_ptr;
-    while (next != 0) {
-	hal_param_t *param = (hal_param_t *)SHMPTR(next);
-	owner = (hal_comp_t *) SHMPTR(param->owner_ptr);
-	if (owner->comp_id == comp->comp_id)
-	    count++;
-	next = param->next_ptr;
-    }
-    return count;
-}
-
-// hal mutex scope-locked version of halpr_find_pin_by_name()
-hal_pin_t *
-hal_find_pin_by_name(const char *name)
-{
-    hal_pin_t *p __attribute__((cleanup(halpr_autorelease_mutex)));
-    rtapi_mutex_get(&(hal_data->mutex));
-    p = halpr_find_pin_by_name(name);
-    return p;
-}
-
-int
-hal_comp_state_by_name(const char *name)
-{
-    hal_comp_t *comp __attribute__((cleanup(halpr_autorelease_mutex)));
-    rtapi_mutex_get(&(hal_data->mutex));
-
-    comp = halpr_find_comp_by_name(name);
-    if (comp == NULL)
-	return -ENOENT;
-    return comp->state;
 }
 
 /***********************************************************************
