+++ src/hal/lib/hal_list.c
@@ -1,0 +1,80 @@
+// HAL doubly-linked list
+
+#include "config.h"
+#include "rtapi.h"		/* RTAPI realtime OS API */
+#include "hal.h"		/* HAL public API decls */
+#include "hal_priv.h"		/* HAL private decls */
+#include "hal_internal.h"
+
+
+hal_list_t *list_prev(hal_list_t * entry)
+{
+    /* this function is only needed because of memory mapping */
+    return SHMPTR(entry->prev);
+}
+
+hal_list_t *list_next(hal_list_t * entry)
+{
+    /* this function is only needed because of memory mapping */
+    return SHMPTR(entry->next);
+}
+
+void list_init_entry(hal_list_t * entry)
+{
+    int entry_n;
+
+    entry_n = SHMOFF(entry);
+    entry->next = entry_n;
+    entry->prev = entry_n;
+}
+
+void list_add_after(hal_list_t * entry, hal_list_t * prev)
+{
+    int entry_n, prev_n, next_n;
+    hal_list_t *next;
+
+    /* messiness needed because of memory mapping */
+    entry_n = SHMOFF(entry);
+    prev_n = SHMOFF(prev);
+    next_n = prev->next;
+    next = SHMPTR(next_n);
+    /* insert the entry */
+    entry->next = next_n;
+    entry->prev = prev_n;
+    prev->next = entry_n;
+    next->prev = entry_n;
+}
+
+void list_add_before(hal_list_t * entry, hal_list_t * next)
+{
+    int entry_n, prev_n, next_n;
+    hal_list_t *prev;
+
+    /* messiness needed because of memory mapping */
+    entry_n = SHMOFF(entry);
+    next_n = SHMOFF(next);
+    prev_n = next->prev;
+    prev = SHMPTR(prev_n);
+    /* insert the entry */
+    entry->next = next_n;
+    entry->prev = prev_n;
+    prev->next = entry_n;
+    next->prev = entry_n;
+}
+
+hal_list_t *list_remove_entry(hal_list_t * entry)
+{
+    int entry_n;
+    hal_list_t *prev, *next;
+
+    /* messiness needed because of memory mapping */
+    entry_n = SHMOFF(entry);
+    prev = SHMPTR(entry->prev);
+    next = SHMPTR(entry->next);
+    /* remove the entry */
+    prev->next = entry->next;
+    next->prev = entry->prev;
+    entry->next = entry_n;
+    entry->prev = entry_n;
+    return next;
+}
