+++ src/hal/lib/hal_vtable.c
@@ -3,6 +3,7 @@
 #include "rtapi.h"		/* RTAPI realtime OS API */
 #include "hal.h"		/* HAL public API decls */
 #include "hal_priv.h"		/* HAL private decls */
+#include "hal_internal.h"
 
 #if defined(ULAPI)
 #include <stdio.h>
